backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.5.0
joblib>=1.3.0
python-multipart==0.0.6
pydantic==2.5.0
requests>=2.31.0
python-dotenv>=1.0.0
